+++ app/src/main/res/layout/fragment_reservation.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.homefragments.reservation.ReservationFragment">
+    tools:context=".presentation.homefragments.reservationList.ReservationListFragment">
 
     <TextView
         android:id="@+id/reservation_fragment_tv"
